1. 给用户管理的登录注册添加了一些代码

+++ src/main/java/com/lanyuan/testspringboot/controller/RoleController.java
@@ -9,7 +9,9 @@ import com.lanyuan.testspringboot.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/role")
@@ -30,14 +32,20 @@ public class RoleController {
 
     @PostMapping("/doAdd")
     public R doAdd( Role u){
-        if(u.getRolename()==null||u.getRolename()==""){
-            return R.error().data("mess","你对象的参数为空");
+        if(u.getRolename()==null || "".equals(u.getRolename())){
+            return R.error().data("mess","角色名称不能为空");
         }else{
-            int i = roleService.addRole(u);
-            if(i>0){
-                return R.ok();
+            Role role = roleService.findByRolename(u.getRolename());
+            System.out.println(role);
+            if(role==null){
+                int i = roleService.addRole(u);
+                if(i>0){
+                    return R.ok();
+                }else{
+                    return R.error();
+                }
             }else{
-                return R.error();
+                return R.error().data("mess","角色已存在");
             }
         }
 
@@ -57,70 +65,125 @@ public class RoleController {
     @DeleteMapping("/doDelRole")
     public R doDelRole(Integer id){
 
-        //先删除原来的关系表
-        roleService.removeRid(id);
-
-        int n = roleService.removeById(id);
-        if(n>0){
-            return R.ok();
+        if(id == null){
+            return R.error().data("mess","id的参数为空");
         }else{
-            return R.error();
+            //先删除原来的关系表
+            roleService.removeRid(id);
+
+            int n = roleService.removeById(id);
+            if(n>0){
+                return R.ok();
+            }else{
+                return R.error();
+            }
         }
     }
 
     @DeleteMapping("/doBathDelRole")
     public R doBathDelRole(Integer[] ids){
 
-        int m = 1;
-
-        for(Integer id : ids){
-            Role role = roleService.getById(id);
-            if(role!=null){
-                m++;
-            }else{
-                m = -1;
-                break;
-            }
-        }
-
-        if(m > 0){
-            for (Integer id : ids){
-                //先删除原来的关系表
-                roleService.removeRid(id);
-            }
-            roleService.doBathDelRole(ids);
-            return R.ok();
+        if(ids == null){
+            return R.error().data("mess","请传入参数,数组不能为空");
         }else{
-            return R.error();
+
+            System.out.println("长度为:"+ids.length);
+
+            if(ids.length == 0){
+                return R.error().data("mess","数组里面的元素不能为空");
+            }
+
+            for (int i = 0; i < ids.length; i++) {
+                if(ids[i] == null){
+                    return R.error().data("mess","id的值不能为空");
+                }
+            }
+            int m = 1;
+
+            for(Integer id : ids){
+                Role role = roleService.getById(id);
+                if(role!=null){
+                    m++;
+                }else{
+                    m = -1;
+                    break;
+                }
+            }
+
+            if(m > 0){
+                for (Integer id : ids){
+                    //先删除原来的关系表
+                    roleService.removeRid(id);
+                }
+                roleService.doBathDelRole(ids);
+                return R.ok();
+            }else{
+                return R.error();
+            }
         }
     }
 
     @GetMapping("/findById")
     public R findById(Integer id){
-        Role role = roleService.getById(id);
-        if(role!=null){
-            return R.ok().data("role",role);
+
+        if(id == null){
+            return R.error().data("mess","id的参数不能为空");
         }else{
-            return R.error();
+            Role role = roleService.getById(id);
+            if(role!=null){
+                return R.ok().data("role",role);
+            }else{
+                return R.error().data("mess","根据此id从数据库查不到该数据");
+            }
         }
     }
 
     @PutMapping("/doUpdateRole")
     public R doUpdateRole(Role role){
-        int i = roleService.updateById(role);
-        if(i>0){
-            return R.ok();
+
+        if(role.getId() == null){
+            return R.error().data("mess","id的参数为空");
         }else{
-            return R.error();
+            Role r = roleService.getById(role.getId());
+            if(r==null){
+                return R.error().data("mess","对象不存在");
+            }else{
+                if(role.getRolename() != null){
+                    if("".equals(role.getRolename())){
+                        return R.error().data("mess","角色名称不能为空");
+                    }else{
+                        int i = roleService.updateById(role);
+                        if(i>0){
+                            return R.ok();
+                        }else{
+                            return R.error();
+                        }
+                    }
+                }else{
+                    return R.error().data("mess","角色名称不能为空");
+                }
+            }
         }
     }
 
     //带条件的分页查询
     @GetMapping("/pageList")
     public R pageList(Role role,@RequestParam(defaultValue = "1") Integer pageNum,
-                          @RequestParam(defaultValue = "5") Integer pageSize){
-        PageInfo<Role> pageInfo = roleService.getPage(role,pageNum,pageSize);
-        return R.ok().data("items",pageInfo.getList());
+                      @RequestParam(defaultValue = "5") Integer pageSize){
+        PageInfo<Role> pageInfo = null;
+        if(role.getRolename()==null || "".equals(role.getRolename())){
+            pageInfo = roleService.getPage(role,pageNum,pageSize);
+            Map<String,Object> map = new HashMap<>();
+            map.put("mess","查询所有的角色");
+            map.put("items",pageInfo.getList());
+            return R.ok().data(map);
+        }else{
+            pageInfo = roleService.getPage(role,pageNum,pageSize);
+            Map<String,Object> map = new HashMap<>();
+            map.put("搜索的条件",role.getRolename());
+            map.put("items",pageInfo.getList());
+            return R.ok().data(map);
+        }
     }
 
     //查询所有的角色
@@ -133,32 +196,55 @@ public class RoleController {
     @PostMapping("/doAddRelation")
     public R doAddRelation(Integer uid,Integer[] rids){
 
-        User user = userService.getById(uid);
-        if(user != null){
-            int m = 1;
+        if(uid == null){
+            return R.error().data("mess","用户id不能为空");
+        }else{
+            if(rids == null){
+                return R.error().data("mess","角色数组不能为空");
+            }else{
 
-            for(Integer id : rids){
-                Role role = roleService.getById(id);
-                if(role!=null){
-                    m++;
-                }else{
-                    m = -1;
-                    break;
+                System.out.println("长度为:" + rids.length);
+
+                if(rids.length == 0){
+                    return R.error().data("mess","角色数组里面的元素不能为空");
                 }
-            }
 
-            if(m > 0){
-                //先删除原来的关系表
-                roleService.removeRelation(uid);
+                for (int i = 0; i < rids.length; i++) {
+                    if(rids[i] == null){
+                        return R.error().data("mess","角色id不能为空");
+                    }
+                }
 
-                //再添加现有的关系
-                roleService.addRelation(uid,rids);
+                User user = userService.getById(uid);
+                if(user != null){
+                    int m = 1;
 
-                return R.ok();
+                    for(Integer id : rids){
+                        Role role = roleService.getById(id);
+                        if(role!=null){
+                            m++;
+                        }else{
+                            m = -1;
+                            break;
+                        }
+                    }
+
+                    if(m > 0){
+                        //先删除原来的关系表
+                        roleService.removeRelation(uid);
+
+                        //再添加现有的关系
+                        roleService.addRelation(uid,rids);
+
+                        return R.ok();
+                    }
+                }
+
+                return R.error();
+
             }
         }
 
-        return R.error();
     }
 
 }
